backend/pyproject.toml
[tool.poetry]
name = "bar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
